further refactor

+++ helper.js
@@ -39,8 +39,13 @@ function bufferToCheerio(buffer) {
   else return tryCatch(() => cheerio.load(buffer.toString()));
 }
 
-function save(filepath, data) {
+function save(filepath, data, dryRun = false) {
   return new Promise((resolve, reject) => {
+    if (dryRun) {
+      callback();
+      return;
+    }
+
     fs.writeFile(filepath, data, 'utf8', err => {
       if (err) {
         console.log('  + error writing file', htmlOutPath, err);
